requirements.txt
ultralytics
pywebview
Pillow
opencv-python
numpy